+++ Changelog.md
@@ -5,6 +5,7 @@ _yyyy.mm.dd_
 
 ### New features
 * Add `Data::PointerWrapper` class to simplify interracing of GEGELATI with primitive variables (non-array) data from a learning environment.
+* Add `TPG::ExecutionsStats` class to analyze and export execution statistics gathered using an instrumented TPGGraph. Statistics include averages on numbers of evaluated teams, programs, program lines and instructions, execution traces and various distributions based on execution traces. The class also provides a method to export these statistics to a JSon file, which can be used by other programs and scripts.
 
 ### Changes
 * CodeGen TPGs (Switch and Stack modes) now match the execution behaviour of `TPGExecutionEngine`, especially regarding cycle handling. Now, a team can't be visited more than once per inference, and edges leading to an already visited team are not evaluated (their program are not executed).